Inline `DataSource.read(…)…`
Now all SQL generation is done by the `DataDomain` itself. Next step is to extract relevant abstractions (such as SQL generation) from `DataDomain`.
